Bug 565917 [Passage] cleanup org.eclipse.passage.lic.json
Remove deprecated JSON transports and its infrastructure
