+++ bundles/org.eclipse.passage.lic.api/META-INF/MANIFEST.MF
@@ -29,11 +29,10 @@ Export-Package: org.eclipse.passage.lic.api,
  org.eclipse.passage.lic.internal.api.conditions;
   x-friends:="org.eclipse.passage.lbc.api,
    org.eclipse.passage.lbc.base,
-   org.eclipse.passage.lic.base",
+   org.eclipse.passage.lic.base,
+   org.eclipse.passage.lic.json",
  org.eclipse.passage.lic.internal.api.conditions.evaluation;x-internal:=true,
- org.eclipse.passage.lic.internal.api.conditions.mining;
-  x-friends:="org.eclipse.passage.lbc.api,
-   org.eclipse.passage.lbc.base",
+ org.eclipse.passage.lic.internal.api.conditions.mining;x-friends:="org.eclipse.passage.lbc.api,org.eclipse.passage.lbc.base,org.eclipse.passage.lic.json",
  org.eclipse.passage.lic.internal.api.diagnostic;x-friends:="org.eclipse.passage.lic.base,org.eclipse.passage.lic.jface",
  org.eclipse.passage.lic.internal.api.inspection;x-internal:=true,
  org.eclipse.passage.lic.internal.api.io;x-friends:="org.eclipse.passage.lbc.api",
