+++ bundles/org.eclipse.passage.lic.base/META-INF/MANIFEST.MF
@@ -16,7 +16,7 @@ Export-Package: org.eclipse.passage.lic.base,
  org.eclipse.passage.lic.base.restrictions,
  org.eclipse.passage.lic.internal.base;x-friends:="org.eclipse.passage.lic.jface,org.eclipse.passage.lbc.base",
  org.eclipse.passage.lic.internal.base.access;x-internal:=true,
- org.eclipse.passage.lic.internal.base.conditions;x-friends:="org.eclipse.passage.lbc.base",
+ org.eclipse.passage.lic.internal.base.conditions;x-friends:="org.eclipse.passage.lbc.base,org.eclipse.passage.lic.json",
  org.eclipse.passage.lic.internal.base.conditions.evaluation;x-internal:=true,
  org.eclipse.passage.lic.internal.base.conditions.mining;x-friends:="org.eclipse.passage.lic.jface",
  org.eclipse.passage.lic.internal.base.diagnostic;x-internal:=true,
