+++ bundles/org.eclipse.passage.lic.json/.settings/.api_filters
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8" standalone="no"?>
+<component id="org.eclipse.passage.lic.json" version="2">
+    <resource path="META-INF/MANIFEST.MF">
+        <filter comment="Bug 565917" id="925892614">
+            <message_arguments>
+                <message_argument value="1.0.0"/>
+                <message_argument value="0.5.200"/>
+            </message_arguments>
+        </filter>
+    </resource>
+</component>
