+++ bundles/org.eclipse.passage.lic.json/META-INF/MANIFEST.MF
@@ -2,7 +2,7 @@ Manifest-Version: 1.0
 Automatic-Module-Name: org.eclipse.passage.lic.json
 Bundle-ManifestVersion: 2
 Bundle-SymbolicName: org.eclipse.passage.lic.json
-Bundle-Version: 0.5.300.qualifier
+Bundle-Version: 1.0.0.qualifier
 Bundle-Name: %Bundle-Name
 Bundle-Vendor: %Bundle-Vendor
 Bundle-Copyright: %Bundle-Copyright
@@ -13,9 +13,7 @@ Require-Bundle: org.apache.httpcomponents.httpcore;bundle-version="0.0.0",
  com.fasterxml.jackson.core.jackson-core;bundle-version="0.0.0",
  com.fasterxml.jackson.core.jackson-databind;bundle-version="0.0.0",
  org.eclipse.osgi.services;bundle-version="0.0.0",
- org.eclipse.passage.lic.base;bundle-version="0.0.0",
+ org.eclipse.passage.lic.base;bundle-version="1.0.0",
  org.eclipse.passage.lic.net;bundle-version="0.0.0"
-Export-Package: org.eclipse.passage.lic.internal.json;x-internal:=true,
- org.eclipse.passage.lic.internal.json.tobemoved;x-internal:=true
+Export-Package: org.eclipse.passage.lic.internal.json;x-internal:=true
 Bundle-ActivationPolicy: lazy
-Service-Component: OSGI-INF/*.xml
