--- bundles/org.eclipse.passage.lic.json/OSGI-INF/org.eclipse.passage.lic.internal.json.JsonConditionTransport.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" name="org.eclipse.passage.lic.internal.json.JsonConditionTransport">
-   <property name="licensing.content.type" value="application/json"/>
-   <implementation class="org.eclipse.passage.lic.internal.json.JsonConditionTransport"/>
-   <service>
-      <provide interface="org.eclipse.passage.lic.api.conditions.ConditionTransport"/>
-   </service>
-</scr:component>
