--- bundles/org.eclipse.passage.lic.json/OSGI-INF/org.eclipse.passage.lic.internal.json.JsonPermissionTransport.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<scr:component xmlns:scr="http://www.osgi.org/xmlns/scr/v1.1.0" name="org.eclipse.passage.lic.internal.json.JsonPermissionTransport">
-   <implementation class="org.eclipse.passage.lic.internal.json.JsonPermissionTransport"/>
-   <property name="licensing.content.type" type="String" value="application/json"/>
-   <service>
-      <provide interface="org.eclipse.passage.lic.api.access.PermissionTransport"/>
-   </service>
-</scr:component>
