--- bundles/org.eclipse.passage.lic.json/src/org/eclipse/passage/lic/internal/json/JsonTransport.java
@@ -1,42 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2018, 2020 ArSysOp
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License 2.0 which is available at
- * https://www.eclipse.org/legal/epl-2.0/.
- *
- * SPDX-License-Identifier: EPL-2.0
- *
- * Contributors:
- *     ArSysOp - initial API and implementation
- *******************************************************************************/
-package org.eclipse.passage.lic.internal.json;
-
-import com.fasterxml.jackson.annotation.JsonAutoDetect;
-import com.fasterxml.jackson.core.JsonParser;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.SerializationFeature;
-
-/**
- * @deprecated use revised version of transport (tobemoved package)
- */
-@Deprecated
-public class JsonTransport {
-
-	private JsonTransport() {
-		// block
-	}
-
-	public static ObjectMapper createObjectMapper() {
-		ObjectMapper mapper = new ObjectMapper();
-		mapper.setVisibility(mapper.getSerializationConfig().getDefaultVisibilityChecker()
-				.withFieldVisibility(JsonAutoDetect.Visibility.ANY)//
-				.withGetterVisibility(JsonAutoDetect.Visibility.NONE)//
-				.withSetterVisibility(JsonAutoDetect.Visibility.ANY)//
-				.withCreatorVisibility(JsonAutoDetect.Visibility.ANY));
-		mapper.enable(SerializationFeature.INDENT_OUTPUT);
-		mapper.enable(JsonParser.Feature.ALLOW_UNQUOTED_FIELD_NAMES);
-		return mapper;
-	}
-
-}
