+++ bundles/org.eclipse.passage.seal.demo/src/org/eclipse/passage/seal/internal/demo/SealedAccessCycleConfiguration.java
@@ -58,7 +58,7 @@ import org.eclipse.passage.lic.internal.equinox.io.BundleKeyKeeper;
 import org.eclipse.passage.lic.internal.equinox.requirements.BundleRequirements;
 import org.eclipse.passage.lic.internal.equinox.requirements.ComponentRequirements;
 import org.eclipse.passage.lic.internal.hc.remote.impl.RemoteConditions;
-import org.eclipse.passage.lic.internal.json.tobemoved.JsonConditionTransport;
+import org.eclipse.passage.lic.internal.json.JsonConditionTransport;
 import org.eclipse.passage.lic.internal.licenses.migration.tobemoved.XmiConditionTransport;
 import org.eclipse.passage.lic.internal.oshi.tobemoved.HardwareAssessmentService;
 import org.eclipse.passage.lic.internal.oshi.tobemoved.HardwareEnvironment;
