+++ features/org.eclipse.passage.lic.json.feature/feature.xml
@@ -14,7 +14,7 @@
 <feature
       id="org.eclipse.passage.lic.json.feature"
       label="%featureName"
-      version="0.6.200.qualifier"
+      version="1.0.0.qualifier"
       provider-name="%providerName"
       plugin="org.eclipse.passage.lic.json"
       license-feature="org.eclipse.license"
