+++ tests/org.eclipse.passage.lic.api.tests/META-INF/MANIFEST.MF
@@ -12,7 +12,7 @@ Require-Bundle: org.junit;bundle-version="4.12.0",
 Export-Package: org.eclipse.passage.lic.api.tests;x-internal:=true,
  org.eclipse.passage.lic.api.tests.conditions;x-internal:=true,
  org.eclipse.passage.lic.api.tests.conditions.evaluation;x-internal:=true,
- org.eclipse.passage.lic.api.tests.conditions.mining;x-internal:=true,
+ org.eclipse.passage.lic.api.tests.conditions.mining;x-friends:="org.eclipse.passage.lic.json.tests",
  org.eclipse.passage.lic.api.tests.fakes.conditions;x-internal:=true,
  org.eclipse.passage.lic.api.tests.fakes.conditions.evaluation,
  org.eclipse.passage.lic.api.tests.fakes.conditions.mining;x-internal:=true,
