+++ tests/org.eclipse.passage.lic.api.tests/src/org/eclipse/passage/lic/api/tests/conditions/mining/ConditionTransportContractTest.java
@@ -28,7 +28,6 @@ import org.eclipse.passage.lic.internal.api.conditions.ValidityPeriodClosed;
 import org.eclipse.passage.lic.internal.api.conditions.mining.ConditionTransport;
 import org.junit.Test;
 
-@SuppressWarnings("restriction")
 public abstract class ConditionTransportContractTest {
 
 	@Test
