+++ tests/org.eclipse.passage.lic.json.tests/META-INF/MANIFEST.MF
@@ -2,17 +2,12 @@ Manifest-Version: 1.0
 Automatic-Module-Name: org.eclipse.passage.lic.json.tests
 Bundle-ManifestVersion: 2
 Bundle-SymbolicName: org.eclipse.passage.lic.json.tests
-Bundle-Version: 0.5.100.qualifier
+Bundle-Version: 1.0.0.qualifier
 Bundle-Name: %Bundle-Name
 Bundle-Vendor: %Bundle-Vendor
 Bundle-Copyright: %Bundle-Copyright
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
+Fragment-Host: org.eclipse.passage.lic.json
 Require-Bundle: org.junit;bundle-version="4.12.0",
- com.fasterxml.jackson.core.jackson-annotations;bundle-version="2.9.2",
- com.fasterxml.jackson.core.jackson-core;bundle-version="2.9.2",
- com.fasterxml.jackson.core.jackson-databind;bundle-version="2.9.2",
- org.eclipse.passage.lic.base;bundle-version="0.0.0",
- org.eclipse.passage.lic.json;bundle-version="0.0.0",
  org.eclipse.passage.lic.api.tests;bundle-version="0.1.0"
-Export-Package: org.eclipse.passage.lic.json.tests,
- org.eclipse.passage.lic.json.tests.tobemoved
+Export-Package: org.eclipse.passage.lic.json.tests
